+++ src/ai_e2e_tester/agent.py
@@ -1,10 +1,12 @@
 import logging
+import os
 import time
 from collections import defaultdict
 from typing import List, Dict
 from urllib.parse import urlparse, urlunparse
 
 from ai_e2e_tester.browser.actions import ACTION_REGISTRY
+from ai_e2e_tester.browser.guards.domain_guard import ensure_stay_on_domain
 from ai_e2e_tester.browser.session import BrowserSession
 from ai_e2e_tester.browser.visited_page import VisitedPage
 from ai_e2e_tester.llm.openai import OpenAiWrapper
@@ -23,8 +25,7 @@ class TestingAgent:
 
         self.url = url
         self.config = load_config(config_path)
-        self.prompt_template = self.config['prompts']['main']
-        self.llm = OpenAiWrapper(self.prompt_template)
+        self.llm = OpenAiWrapper(prompts=self.config['prompts'])
 
         self.reporter = ConsoleReporter()
 
@@ -35,37 +36,39 @@ class TestingAgent:
     def run(self, max_steps: int):
         logger.info(f"Test started. Max steps: {max_steps}")
         browser_session = BrowserSession(self.url, headless=True)
-        browser_session.goto_url(self.url)
+
+        goal = None
+        expectation = None
 
         for step_idx in range(max_steps):
 
-            # Guard BEFORE LLM
-            # @todo Inform LLM that we navigated back after leaving domain.
-            if not browser_session.ensure_stay_on_domain():
-                break
-
             logger.info(f'[Step {step_idx + 1}] On Page: {browser_session.url}')
 
-            text = browser_session.get_page_html()
+            user_prompt = self._get_user_prompt(browser_session)
+            system_prompt = self._get_system_prompt()
             screenshot_b64 = browser_session.get_screenshot(path=f"reports/screenshot_{step_idx + 1}.png")
 
-            result = self.llm.run(
-                page_url=browser_session.url,
-                page_html=text,
-                screenshot_b64=screenshot_b64,
-                context=self._generate_llm_context(),
-                available_actions=self._generate_llm_available_actions()
-            )
+            self._save_report(f"prompt-{step_idx + 1}.txt", user_prompt)
+
+            result = self.llm.run(system_prompt, user_prompt, screenshot_b64)
+
+            goal = result.get('updated_goal') or result.get('goal')
+            expectation = result.get('expected_vs_actual')
+            logger.info(f'Current goal: {goal}')
+            logger.debug(f'Current expectations: {expectation}')
 
             visited_page = VisitedPage.from_json(browser_session.page, result)
-            logger.info(f'{visited_page.summary}')
             self.visited_pages.append(visited_page)
 
             if visited_page.has_next_step():
+                browser_session.clear_console_messages()
                 visited_page.run_next_step(browser_session)
+                ensure_stay_on_domain(browser_session, visited_page)
+                logger.info(visited_page.next_step.get_feedback_summary())
             else:
                 logger.info("The LLM has decided that there is nothing more to do.")
                 break
+
             time.sleep(self.wait_between_steps)
 
         browser_session.close()
@@ -82,7 +85,8 @@ class TestingAgent:
         grouped_visits = defaultdict(list)
         for page in self.visited_pages:
             parts = urlparse(page.page_url)
-            clean_url = urlunparse(parts._replace(fragment=''))
+            normalized = parts._replace(query='', fragment='')
+            clean_url = urlunparse(normalized)
             grouped_visits[clean_url].append(page)
         return grouped_visits
 
@@ -98,3 +102,21 @@ class TestingAgent:
         return "\n".join(
             action_cls.describe_for_llm() for action_cls in ACTION_REGISTRY.values()
         )
+
+    @classmethod
+    def _save_report(cls, name: str, content: str):
+        os.makedirs('reports', exist_ok=True)
+        with open(f"reports/{name}", "w", encoding="utf-8") as f:
+            f.write(content)
+
+    def _get_user_prompt(self, browser_session: BrowserSession) -> str:
+        return self.config['prompts']['user'].format(
+            page_url=browser_session.url,
+            page_html=browser_session.get_optimized_html(),
+            console_logs=browser_session.get_console_messages(),
+            context=self._generate_llm_context(),
+            available_actions=self._generate_llm_available_actions()
+        )
+
+    def _get_system_prompt(self) -> str:
+        return self.config['prompts']['system']