+++ CHANGELOG.md
@@ -1,7 +1,12 @@
+### 0.0.3
 
-### 1.0.1
+- Improved feedback quality and reduced duplicated feedback.
+- Optimized HTML passed to the LLM to reduce token usage.
+- Better location of elements on the page based on LLM description.
+
+### 0.0.2
 
 - Description of available browser actions for the LLM is now automatically generated from the list of declared action classes.
 
-### 1.0.0
+### 0.0.1
 First basic version.